+++ frontend/app/src/DevApp.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { Routes } from '@/router/routes';
+import { useHistoryEventCounterpartyMappings } from '@/composables/history/events/mapping/counterparty';
 
 function reset() {
   sessionStorage.removeItem('vuex');
@@ -7,6 +8,53 @@ function reset() {
 }
 
 const resetState = 'Reset State';
+
+const { supportedChains } = useSupportedChains();
+const { allLocations, allExchanges } = storeToRefs(useLocationStore());
+const { getCounterpartyData, counterparties } = useHistoryEventCounterpartyMappings();
+
+const imageUrl = 'https://raw.githubusercontent.com/rotki/rotki/develop/frontend/app/public/assets/images/protocols/';
+
+const integrationData = computed(() => {
+  const blockchains = get(supportedChains)
+    .filter(item => item.name !== 'Ethereum Staking')
+    .map(item => ({
+      label: toSentenceCase(item.name),
+      image: `${imageUrl}${item.image}`,
+    }));
+
+  const exchanges = get(allExchanges).map((item) => {
+    const data = get(allLocations)[item];
+
+    return {
+      label: data.label || toSentenceCase(item),
+      image: `${imageUrl}${data.image}`,
+    };
+  });
+
+  const protocols = get(counterparties)
+    .filter(item => item !== 'gas' && !(/v\d+$/).test(item)) // remove gas, and anything that has version number
+    .filter(uniqueStrings)
+    .map((item) => {
+      const data = get(getCounterpartyData(item));
+      return {
+        label: toHumanReadable(data.label, 'sentence'),
+        image: `${imageUrl}${data.image}`,
+      };
+    });
+
+  const data = {
+    blockchains,
+    exchanges,
+    protocols,
+  };
+
+  return JSON.stringify(data, null, 2);
+});
+
+const { copy: copyGeneratedIntegrationData } = useClipboard({ source: integrationData });
+
+const copyText = 'Copy Integration JSON Data';
 </script>
 
 <template>
@@ -26,6 +74,12 @@ const resetState = 'Reset State';
         </RuiButton>
       </RouterLink>
       <div class="grow" />
+      <RuiButton
+        color="primary"
+        @click="copyGeneratedIntegrationData()"
+      >
+        {{ copyText }}
+      </RuiButton>
       <RuiButton
         variant="outlined"
         color="warning"
